requirements.txt
bcrypt==3.2.0
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dotenv==0.19.0
SQLAlchemy==1.4.25

